refactor: begin separating BlogTopiContext into proper file structure
