+++ frontend/src/context/blog_form_context/createBlogFormContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { BlogFormContextType } from "../BlogFormContext";
+
+export const BlogFormContext = createContext<BlogFormContextType | undefined>(
+  undefined
+);
