+++ frontend/src/features/generate/components/BlogTopicForm.tsx
@@ -1,6 +1,6 @@
 import { Card } from "../../../components/Card";
 import { FaTimes } from "react-icons/fa";
-import { useBlogForm } from "../../../context/BlogFormContext";
+import { useBlogForm } from "../../../context/useBlogFormContext";
 import { useEffect, useState } from "react";
 
 /**
